packages/model/pyproject.toml
[project]
name = "skin-triage-model"
version = "0.1.0"
description = "Skin triage model package"
authors = [
    "Clinikly"
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "skin-triage-model"
version = "0.1.0"
description = "Skin triage model package"
authors = ["Clinikly"]
readme = "README.md"
packages = [{include = "skin_triage_model"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = ">=2.0.0,<3.0.0"
pillow = ">=11.2.1,<12.0.0"
torchvision = ">=0.15.0,<0.19.0"
fastapi = ">=0.100.0,<1.0.0"
uvicorn = {extras = ["standard"], version = ">=0.22.0,<1.0.0"}
